Port diff functions

+++ src/main/kotlin/com/fitness/fitness_backend/entity/diff/Diff.kt
@@ -3,6 +3,6 @@ package com.fitness.fitness_backend.entity.diff
 import com.fitness.fitness_backend.entity.Client
 
 data class Diff(
-    var timestamp: Int,
+    var timestamp: Long,
     var clients: List<Client>
 )
